+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Changed
+- Created `KeyedTableDiff.Builder`.
+
+### Removed
+- Removed public constructors of `KeyedTableDiff`. Use `KeyedTableDiff.Builder`.
+
+### Fixed
+- Ignore empty rows in `KeyedTableDiff`. #36
+
+
 ## [0.25.1] - 2022-10-15
 ### Changed
 - `RowDiff` now accept more data in a row than declared in the corresponding header.